guide-slack-chart-patch/requirements_additions.txt
# Additional requirements for chart support
vl-convert-python>=1.0.0
